Split resolver symbol table storage
